Do not refresh the auth token on initialize
Delay the fetching of the auth token to when the connection to the schema registry is used.

This is to prevent making an HTTP call during initialization of AvroTurf, which, if done as part of a Rails initializer would have a negative impact on startup performance of a Rails app.