Expose messaging.schema_store and messaging.registry
This provides accessors for the `schema_store` and `registry`.

Allowing read + write access to the schema store to allow dynamic setting of a schema store with a different path. And to trigger loading of schemas on demand.
Allowing read access to the registry so that we can do things like check compatibility (which is exposed on the registry class).